refactor: extract AuthPromoPanel out of Login.tsx
Shared by Login and Register (next commits) instead of duplicated.
